Starting to work on level design for the first zone
Most of my work thus far has been setting up the code and UI to have
systems in place. Now I am using them to create levels. It will be
possible to add new things in the future though.

Added a bunch of decorative prefabs as well, and modified some elevator
scripts to function correctly when pressed or shot at.