Big update! Persistent Game Manager.
Developed a system using scripts and game objects which will essentially
allow the player to save their game status (which levels they have
beaten, and if they are currently in a level, which half-way mark they
have accessed). Once the player dies in a level, using the new scripts
the player will teleport to their most recent save-point within the
level.